Super Scope Example update
Calibration now works as it should. scope_held functions are best explained too. Offscreen sprites now are hidden. Removed some unnecessary messages because we don't know if the fist aim adjust shot is in the middle of screen, so it's just a test screen.

Noise and Offscreen bits are not used in this example, but they are present on lib.