Fix scanMplay5
The getpad1data loop did a 16-bit read of REG_JOYA, which also reads
from REG_JOYB, preventing the getpad23data loop from reading joypad
values.

scanMplay5 now tests if a standard controller is connected to the
multitap.

This commit also optimises the multitap reading code by:
 * Using Joypad Auto-Read to read the first three controllers.
 * Using ScanPads to read the first two controllers.
 * Using an 8 bit A when reading the last two controllers.
 * Using the carry output of `rol` to determine when all of the
   controller bits have been read.

As a consequence of the Joypad Auto-Read optimisation, the WRIO register
must not be written to when `snes_mplay5` is set.  I do not think this
will be an issue, since `REG_WRIO` (nor 4201) is not defined in the
header files nor the snes-examples directory.

Tested by running the multiplay5 example in Mesen