Fix potential nmiSet() crash
Writes to nmi_handler are not atomic and the VBlank ISR can crash if a
VBlank Interrupt occurs in the middle of the nmi_handler write.

To fix this crash, nmiSet() is now a function that will temporarily
disable interrupts so nmi_handler can be safely written to.

This commit also modifies consoleInit() to use nmiSet().