Do not clear snes_mouse if the mouse has been disconnected
If the developer was unaware `snes_mouse` was cleared by the VBlank ISR
then the player could lock the game if the mouse was ever disconnected.

`snes_mouse` can no-longer be used to detect if the mouse has been
disconnected from the console.  The developer will have to check the
`mouseConnect[]` variable instead.