Delay mouse sensitivity changes to the VBlank ISR (changing the Mouse API)
The old `mouseSpeedChange(u8 port)` function has a timing conflict with
the auto-joypad-read.  The automatic joypad read must not start while
the CPU is in the middle of the `mouseSpeedChange()` function.

Delaying all mouse sensitivity changes to the VBlank ISR will remove
this timing conflict and ensure the mouse sensitivity is cycled at the
correct time.

This commit removes the old `mouseSpeedSet` variable and the
`mouseSpeedChange()` function, replacing them with a
`mouseRequestChangeSensitivity` variable that will signal to
`_MouseData` that the MainLoop wants to either:
 * Cycle the mouse sensitivity once.
 * Cycle the mouse sensitivity twice (decrementing the sensitivity).
 * Setting the mouse sensitivity to a specific value.

This commit also adds a `mouseSensitivity` variable that serves three
purposes:
 1. Allows the developer to read the reported mouse sensitivity.
 2. Allows the developer to set the initial mouse sensitivity.
 3. Restore the mouse sensitivity when the mouse is disconnected, then
    reconnected.

Three helper functions have been added to simplify the
`mouseRequestChangeSensitivity` writes:
 * void mouseCycleSensitivity(u16 port)
 * void mouseCycleSensitivityTwice(u16 port)
 * void mouseSetSensitivity(u16 port, u16 sensitivity)

The two mouse examples have been modified to use the new mouse
sensitivity functions and to display the reported mouse sensitivity.